% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_api.R
\name{mod_api_server}
\alias{mod_api_server}
\title{API Data tab module server}
\usage{
mod_api_server(id, raw_usage_rv)
}
\arguments{
\item{id}{Shiny module ID.}

\item{raw_usage_rv}{`reactiveVal` holding the full usage tibble. API data
can be merged into this value when the user confirms.}
}
\description{
API Data tab module server
}
