% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data.R
\name{mod_data_server}
\alias{mod_data_server}
\title{Data tab module server}
\usage{
mod_data_server(id, raw_usage_rv)
}
\arguments{
\item{id}{Shiny module ID.}

\item{raw_usage_rv}{`reactiveVal` holding the full usage tibble.}
}
\description{
Data tab module server
}
