% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_model.R
\name{mod_model_server}
\alias{mod_model_server}
\title{By Model tab module server}
\usage{
mod_model_server(id, filtered_rv)
}
\arguments{
\item{id}{Shiny module ID.}

\item{filtered_rv}{Reactive tibble of filtered usage data (from
[mod_overview_server()]).}
}
\description{
By Model tab module server
}
