% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_project.R
\name{mod_project_server}
\alias{mod_project_server}
\title{By Project tab module server}
\usage{
mod_project_server(id, filtered_rv)
}
\arguments{
\item{id}{Shiny module ID.}

\item{filtered_rv}{Reactive tibble of filtered usage data.}
}
\description{
By Project tab module server
}
