% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sessions.R
\name{mod_sessions_server}
\alias{mod_sessions_server}
\title{Sessions tab module server}
\usage{
mod_sessions_server(id, filtered_rv)
}
\arguments{
\item{id}{Shiny module ID.}

\item{filtered_rv}{Reactive tibble of filtered usage data.}
}
\description{
Sessions tab module server
}
