% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_upload.R
\name{mod_upload_server}
\alias{mod_upload_server}
\title{Upload Data tab module server}
\usage{
mod_upload_server(id, raw_usage_rv)
}
\arguments{
\item{id}{Shiny module ID.}

\item{raw_usage_rv}{`reactiveVal` holding the full usage tibble. Uploaded
data is merged into this value.}
}
\description{
Upload Data tab module server
}
