% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_data.R
\name{util_empty_usage_tbl}
\alias{util_empty_usage_tbl}
\title{Return an empty usage tibble with the canonical schema}
\usage{
util_empty_usage_tbl()
}
\value{
Zero-row tibble with correct column types.
}
\description{
Return an empty usage tibble with the canonical schema
}
